sw_services: xilplmi: Fixes for MISRA C violation Rule 10.3
MISRA C Rule 10.3 States: The value of an expression shall not be assigned to an
object with a narrower essential type or of a different essential type category

This patch fixes MISRA C Rule 10.3 violation by using the proper type variables
where required or typecasting if necessary


Acked-for-series: Vikram Sreenivasa Batchali <b.vikram@xilinx.com>

+++ lib/sw_services/xilplmi/src/xplmi.c
@@ -154,9 +154,9 @@ void XPlm_PrintPlmBanner(void)
 	u32 Version;
 	u32 PsVersion;
 	u32 PmcVersion;
-	static u8 IsBannerPrinted = FALSE;
+	static u8 IsBannerPrinted = (u8)FALSE;
 
-	if (FALSE == IsBannerPrinted) {
+	if ((u8)FALSE == IsBannerPrinted) {
 		/* Print the PLM Banner */
 		XPlmi_Printf(DEBUG_PRINT_ALWAYS,
                  "****************************************\n\r");
@@ -184,7 +184,7 @@ void XPlm_PrintPlmBanner(void)
 #endif
 		XPlmi_Printf(DEBUG_PRINT_ALWAYS,
                  "****************************************\n\r");
-		IsBannerPrinted = TRUE;
+		IsBannerPrinted = (u8)TRUE;
 	}
 }
